Fix dashboard question details
[#166414256]
