+++ test/components/dashboard/answers-table_spec.js
@@ -1,8 +1,7 @@
 import React from "react";
 import { fromJS } from "immutable";
-import { Provider } from "react-redux";
 import AnswersTable from "../../../js/containers/report/answers-table";
-import { mount } from "enzyme";
+import { mountWithStore } from "../../setupTest";
 
 describe("<AnswersTable />", () => {
   const question = fromJS({
@@ -11,7 +10,6 @@ describe("<AnswersTable />", () => {
   });
   const hidden = false;
   const showCompare = false;
-  const answers = fromJS([]);
   const state = fromJS({
     report: {
       anonymous: false,
@@ -19,29 +17,21 @@ describe("<AnswersTable />", () => {
         1: { id: 1, name: "John Doe" },
         2: { id: 2, name: "Test Student" }
       },
+      answers: {},
       feedbacks: {}
     }
   });
-  const store = { getState: () => state, subscribe: () => {}, dispatch: () => {} };
-  const params = { hidden, showCompare, question, answers };
+  const params = { hidden, showCompare, question };
 
   it("should render student names", () => {
-    const wrapper = mount(
-      <Provider store={store} >
-        <AnswersTable {...params} />
-      </Provider>
-    );
+    const wrapper = mountWithStore(<AnswersTable {...params} />, state);
     expect(wrapper.text()).toEqual(expect.stringContaining("John Doe"));
     expect(wrapper.text()).toEqual(expect.stringContaining("Test Student"));
   });
 
   describe("with a question", () => {
     it("should render <AnswersTable> with Score and Feedback text", () => {
-      const wrapper = mount(
-        <Provider store={store} >
-          <AnswersTable {...params} />
-        </Provider>
-      );
+      const wrapper = mountWithStore(<AnswersTable {...params} />, state);
       expect(wrapper.text()).toMatch(/Student/);
       expect(wrapper.text()).toMatch(/Response/);
       expect(wrapper.text()).toMatch(/Score/);
@@ -54,11 +44,7 @@ describe("<AnswersTable />", () => {
     const params = { hidden, showCompare, question };
 
     it("should render <AnswersTable> without Score or Feedback text", () => {
-      const wrapper = mount(
-        <Provider store={store} >
-          <AnswersTable {...params} />
-        </Provider>
-      );
+      const wrapper = mountWithStore(<AnswersTable {...params} />, state);
       expect(wrapper.text()).toMatch(/Student/);
       expect(wrapper.text()).toMatch(/Response/);
       expect(wrapper.text()).not.toMatch(/Score/);
